backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain and LangGraph
langchain==0.2.16
langgraph==0.2.28
langchain-core==0.2.38
langchain-community==0.2.16
langchain-huggingface==0.0.3
langsmith==0.1.117

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1

# Data processing
jsonlines==4.0.0

# Google Generative AI (if using Gemini)
langchain-google-genai==1.0.10
google-generativeai==0.3.2